Extract db.createCollection to define file
Cut-paste, no code has changed
